requirements.txt
Django~=2.2.5
django-bootstrap-datepicker-plus~=3.0.5
django-bootstrap3~=11.1.0
django-jquery~=3.1.0
Pillow~=6.2.1
